refactor: move pkg/response into core/module/handler
pkg/response had exactly one consumer (stdHandler.go) and was never a
shared library — it was application code accidentally placed in pkg/.
The ACK/NACK body construction was also split between pkg/response and
core/module/handler/responsestep.go (buildAckBody), creating an
artificial boundary.

Changes:
- Delete pkg/response/ entirely
- Add core/module/handler/response.go with all functions unexported
  (sendAck, sendNack, nackBytes, nack, nackBecknError, nackBodyBytes,
  isAtLeastV2, msgID, internalServerError, preV2* types)
- Migrate pkg/response/response_test.go → core/module/handler/response_test.go
- Remove the pkg/response import from stdHandler.go; call sites use the
  now-inlined helpers directly

No behaviour change — same wire format, same error mappings, same tests.
